Allow any arbitrary frequency to be specified for ANU devices and try to find the nearest frequency when initialising it, reporting if the frequency is not exactly as requested.

+++ miner.h
@@ -968,7 +968,7 @@ extern bool opt_restart;
 #ifdef USE_ICARUS
 extern char *opt_icarus_options;
 extern char *opt_icarus_timing;
-extern int opt_anu_freq;
+extern float opt_anu_freq;
 #endif
 extern bool opt_worktime;
 #ifdef USE_AVALON
